Updating user profile pic

+++ Backend/src/main/java/tech/mineapp/controller/FileUploadController.java
@@ -12,10 +12,12 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import tech.mineapp.entity.UserEntity;
 import tech.mineapp.model.response.ContainerResponseModel;
 import tech.mineapp.security.CurrentUser;
 import tech.mineapp.security.UserPrincipal;
 import tech.mineapp.service.FileUploadService;
+import tech.mineapp.service.UserService;
 
 /**
  * @author utkarsh
@@ -25,12 +27,16 @@ import tech.mineapp.service.FileUploadService;
 @RestController
 public class FileUploadController {
 	
-	private final FileUploadService fileUploadService;
+	@Autowired
+	private FileUploadService fileUploadService;
 	
 	@Autowired
-	public FileUploadController(FileUploadService fileUploadService) {
-		this.fileUploadService = fileUploadService;
-	}
+	private UserService userService;
+	
+//	@Autowired
+//	public FileUploadController(FileUploadService fileUploadService) {
+//		this.fileUploadService = fileUploadService;
+//	}
 	
 	@GetMapping("/user/me/pic")
 	@PreAuthorize("hasRole('USER')")
@@ -50,8 +56,20 @@ public class FileUploadController {
 		response.setEndpoint("/user/me/password");
 		
 		try {
-			response.setStatus("SUCCESS");
+			if (!userService.checkVerificationByUserId(userPrincipal.getUserId())) {
+	     		response.setStatus("FAIL");
+	     		response.setErrorMessage("Unverified user.");
+	     		return ResponseEntity.badRequest().body(response);
+	     	}
+			UserEntity user = userService.findUserById(userPrincipal.getUserId());
+			if (!userService.isLocalUser(user)) {
+				response.setStatus("FAIL");
+	     		response.setErrorMessage("Not a local user.");
+	     		return ResponseEntity.badRequest().body(response);
+			}
 			fileUploadService.store(file, String.valueOf(userPrincipal.getUserId()));
+			userService.updateUserProfilePic(user, "http://api.mineapp.tech/user/me/pic");
+			response.setStatus("SUCCESS");
 			
 			return ResponseEntity.ok(response);
 		} catch (Exception e) {
@@ -72,8 +90,21 @@ public class FileUploadController {
 		response.setEndpoint("/user/me/password");
 		
 		try {
-			response.setStatus("SUCCESS");
+			if (!userService.checkVerificationByUserId(userPrincipal.getUserId())) {
+	     		response.setStatus("FAIL");
+	     		response.setErrorMessage("Unverified user.");
+	     		return ResponseEntity.badRequest().body(response);
+	     	}
+			UserEntity user = userService.findUserById(userPrincipal.getUserId());
+			if (!userService.isLocalUser(user)) {
+				response.setStatus("FAIL");
+	     		response.setErrorMessage("Not a local user.");
+	     		return ResponseEntity.badRequest().body(response);
+			}
 			fileUploadService.delete(String.valueOf(userPrincipal.getUserId()));
+			userService.updateUserProfilePic(user, null);
+			
+			response.setStatus("SUCCESS");
 			
 			return ResponseEntity.ok(response);
 		} catch (Exception e) {
